feat: cache fetched content claims in Locator
Add another caching `Map` in `ContentClaimsLocator`. This one caches the fetched content claims and gives callers a way to access the claims that were used as the basis for the `Location` returned by `locate`.

I'm not totally convinced this is the right way to handle this, but it does give downstream users the ability to get the raw content claims associated with a CID.